clapmintic/backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.1
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
gunicorn==20.1.0
django-heroku==0.3.1
